Implemented cards feature
Implemented the Deck class, hand class and card class with necessary methods.
For the Deck, a deck configuration was created so you can easily modify the number of cards in the deck.
Deck:: Implemented the draw, and return card to deck methods
Hand:: Implemented the add card, and use card in hand methods
Card:: Implemented the play card method